requirements.txt
dgl
dgllife
torch
rdkit
pandas
scikit-learn
PyYAML
packaging